Switch tools to non-dev-dependencies installed via phive.
Because filling up the dev-dependencies with normally independent
tools has only benefits in disk size and leads to potentially incomplete
versions in the dependency tree it is advised to reduce the
dev-dependencies as much as possible.

So this commit is it: switch three critical dev tools to their phar
binary distribution.